create_data_dictionary() outputs `tyepof` column
Before the type of each column was manually included in the
`definition` of each `column`. This has two problems:
  1. Does not follow the tidy data principle by which each column
     should map to a single variable (instead it maps to two
     variables, `definition` and `typeof`).
  2. Is error prone because the type was manually entered.

To solve these problems, create_data_dictionary() now outputs the
new `typeof` column which automatically detects the typeof() each
column.

This sets the road for a new approach to defining each column of
contributed datasets.