Define dataset in data-raw/data_dictionary.R
Closes #6.

The `definition` of each `column` is now directly exposed on
data-raw/data_dictionary.R (not data_dictionary_*.csv files).

Benefits:

* This makes the creation of `data_dictionary` more transparent
because the text is right in front of your eyeballs; you
don't even need to read a .csv file.

* Also this makes it easier to track changes to
data_dictionary, because git diff will clearly show any
change to the definition of type of each column.

* It make the data-raw/ directory simpler, as we no longer
need a data_dictionary_*.csv file to define each new dataset.

Downsides:

I expect some tension because people is used to handling this
think of data in Excel (or .csv) files. I'm happy to consider
alternatives if they are similarly simple or a bit more
complex but much easier for the potential contributors (so
that it takes work off the shoulders of the maintainers).

Also:

* No longer ignore references for regression tests:

Before we simply ignored all regression references by matching
"ref*". This was a handy way to ensure no private data was
exposed to the public. But it also ignored public datasets such
as data_dictionary. We now solve this problem by negating the
pattern that would otherwise be ignored.

* Polish documentation:

@param `data` more closely follows the tidyverse principles
that guide this project (https://principles.tidyverse.org).

Thanks @jhoff-2dii